Rename auth.js to oAuthRoutes.js
